+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - post-processing for episodes in FeedCompositionJob
 - `wildfire.calfire` provider
 - `proper_name` and `region` for `wildfire.calfire` provider
+- `wildfire.perimeters.nifc` and `wildfire.locations.nifc` providers
 
 ### Changed